--- LunkGame/game/settings.py
@@ -1,20 +1,0 @@
-"""
-Sean Nishi
-Settings/conststants for Lunk Game
-11/21/2022
-"""
-
-#Screen dimensions
-WINDOW_SIZE = (1280, 720)
-WINDOW_WIDTH = WINDOW_SIZE[0]
-WINDOW_HEIGHT = WINDOW_SIZE[1]
-WINDOW_CENTER = (WINDOW_WIDTH/2, WINDOW_HEIGHT/2)
-
-FPS = 60
-TILESIZE = 64
-
-#image paths
-
-GAME_ICON_PATH = 'graphics/game_icon.jpg'
-MAIN_MENU_BACKGROUND_PATH = 'graphics/sad_start_screen.png'
-
