--- LunkGame/game/support.py
@@ -1,27 +1,0 @@
-#Sean Nishi
-#Lunk Game import_support
-#created 1/24/2023
-
-from csv import reader
-from os import walk
-import pygame
-
-def import_csv_layout(path):
-    terrain_map = []
-    with open(path) as level_map:
-        layout = reader(level_map, delimiter=',')
-        for row in layout:
-            terrain_map.append(list(row))
-        return terrain_map
-
-
-def import_folder(path):
-    surface_list = []
-    #import all images from a folder
-    for _, __, image_files in walk(path):
-        for image in image_files:
-            full_path = path + '/' + image
-            image_surface = pygame.image.load(full_path).convert_alpha()
-            surface_list.append(image_surface)
-
-#import_folder('../graphics/wall')
