--- LunkGame/levels/level_data/wall.tsx
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<tileset version="1.9" tiledversion="1.9.2" name="wall" tilewidth="64" tileheight="64" tilecount="1" columns="1">
- <image source="../../graphics/wall/wall.png" width="64" height="64"/>
-</tileset>
