--- LunkGame/levels/tile_sets/terrain_tiles.tsx
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<tileset version="1.9" tiledversion="1.9.2" name="tile" tilewidth="64" tileheight="64" tilecount="1" columns="1">
- <image source="../../graphics/tile.png" width="64" height="64"/>
-</tileset>
