+++ game/__pycache__/enemy1.cpython-38.pyc
