+++ graphics/floor_surface/ground.png
