+++ graphics/floor_tile/tile.png
