+++ graphics/plant1/plant1.png
