+++ graphics/plant2/plant2.png
