+++ graphics/wall/wall-jungle.png
