+++ graphics/wall/wall-lava.png
